Move helper function to separate file
